netlist-vis-backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "netlist-vis-backend"
version = "0.1.0"
description = "FastAPI backend for the netlist visualizer application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastapi", "pydantic[email]", "python-dotenv", "motor", "uvicorn"]

[project.optional-dependencies]
dev = [
  "fastapi-cli",
  "httpx",
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings",
  "mkdocstrings-python",
  "pytest",
  "pytest-asyncio",
  "requests",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_fixture_scope = "function"
